Features and bug fixes:
Feature this will handle a case where the combinations of generated letters can create 20+ words.
Ideally we want this not to exceed 20 words so in this case we regenerate a new set of letters, then find all possible words.
this is iterated as many times necessary.

Bug fix:
add cleanup to curOffset after iterating through dictionary.